Support for matchers in Internet Explorer 9.
IE9 supports Object.defineProperty but sadly has a bug with number
primitives, where this in a getter would always be 0, which makes
matchers useless when working with numbers in IE9.

But! We can (almost) safely extend Number.prototype with a method,
and thus use that method to overcome the above limitation.

And it works.