First pass at spies.
Spies are capable to create a new function, using `spy(name)`, or on
an object, using `obj.spy(name)` instead. The spies can be verified
to have been called, using `verify()`, and that it was called with a
given list of arguments, using `verify(arg1, arg2)`. Arguments can
also be loosely specified, like for mocks, using `verify(Number,
Object)` for example.

There are no assertions nor expectations for spies yet.